rerfactorf: 수요자 마이페이지 관련 수정
