+++ src/apis/consumer/index.ts
@@ -6,6 +6,7 @@ import type {
   LikedManagerResponse,
   BlackListedManagerResponse,
   PreferenceRequest,
+  ConsumerProfileCreateRequest,
 } from '@/types/consumer';
 import { API_ENDPOINTS } from '@/constants/api';
 
@@ -19,6 +20,13 @@ export const consumerApi = {
       API_ENDPOINTS.CONSUMER.PROFILE,
     );
   },
+
+  /**
+   * 프로필 생성
+   */
+  createProfile: async (data: ConsumerProfileCreateRequest): Promise<void> => {
+    return apiCall<void>('post', API_ENDPOINTS.CONSUMER.PROFILE, data);
+  }, 
 
   /**
    * 프로필 수정
@@ -59,12 +67,12 @@ export const consumerApi = {
   },
 
   /**
-   * 찜한 매니저 제거
+   * 찜한 매니저 삭제
    */
   removePreferenceManager: async (managerUuid: string): Promise<void> => {
     return apiCall<void>(
       'delete',
-      API_ENDPOINTS.CONSUMER.REMOVE_PREFERENCE_MANAGER(managerUuid),
+      API_ENDPOINTS.CONSUMER.REMOVE_PREFERENCE(managerUuid),
     );
   },
 
@@ -77,4 +85,6 @@ export const consumerApi = {
       API_ENDPOINTS.CONSUMER.BLACKLIST,
     );
   },
+
+  
 };
